Move dealias to utilities/internal
